Only firmware files with a valid footer are accepted
This version of the bootloader verifies the firmware files to guard
against data corruption.

Each firmware file should end with
a 64-byte footer with the following format:
- (11 bytes) ASCII characters "3853:sha256" as an indicator
    that a valid 64-byte footer exists.
- (21 bytes) reserved for future use
- (32 bytes) sha256 hash of the firmware file excluding the 64-byte
  footer.

If a firmware file does not have the footer or the sha256 hash in the
footer does not match the content, the file is not written to flash.

If no valid firmware file is found, the bootloader assumes the
flash to contain a valid previous firmware. A future version of the
bootloader may choose not to boot unless the flash has a valid firmware
with a matching footer.